+++ src/js/content/cellar.js
@@ -2,21 +2,25 @@ content.cellar = (() => {
   return {
     export: function () {
       return {
+        deaths: this.deaths.export(),
         discovered: this.discovered.export(),
         health: this.health.export(),
         instruments: this.instruments.export(),
         position: this.position.export(),
         run: this.run.export(),
         scans: this.scans.export(),
+        stockroom: this.stockroom.export(),
       }
     },
     import: function (data = {}) {
+      this.deaths.import(data.deaths)
       this.discovered.import(data.discovered)
       this.health.import(data.health)
       this.instruments.import(data.instruments)
       this.position.import(data.position)
       this.run.import(data.run)
       this.scans.import(data.scans)
+      this.stockroom.import(data.stockroom)
 
       return this
     },
@@ -32,11 +36,13 @@ content.cellar = (() => {
       return health >= 1
     },
     reset: function () {
+      this.deaths.reset()
       this.discovered.reset()
       this.health.reset()
       this.instruments.reset()
       this.run.reset()
       this.scans.reset()
+      this.stockroom.reset()
 
       return this
     },
@@ -47,6 +53,7 @@ content.cellar = (() => {
       this.discovered.reset()
       this.position.reset()
       this.scans.reset()
+      this.stockroom.reset().generate()
       this.tiles.reset()
 
       this.health.setMax()